Improve kernel provider to fetch kernel from network
